move files and add documentation
